+++ components/drivers/virtio/virtio.h
@@ -18,6 +18,7 @@
 #ifdef RT_USING_SMART
 #include <mmu.h>
 #include <ioremap.h>
+#include <lwp_user_mm.h>
 #endif
 
 #if RT_NAME_MAX < 16
